website: move all algolia stuff to env
None of this stuff is useful to be shared, so I've moved it all to just
read from env and updated the deploy-docs GHA environment to include
those vars.

Also set the deploy-docs workflow to support workflow_dispatch so we can
manually trigger a build in case those vars change, for whatever reason.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>