Don't unlazy builds
When we evaluate the state this breaks the debugger.

There is still one place where we unlazy the build, and that is in the
test runner when we need to read a file off the file system for the test
results.
This is a larger change and will be hanlded in a separate commit.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>