spec tests: Add ability to check symlink target
This adds 2 new check options and changes the default back to following
symlinks.

1. Check the target of a symlink (link_target)
2. Do not follow symlinks during checks (no_follow)

`no_follow` is useful if you want to validate just the symlink itself.
Without it, if a symlink is broken (target path does not exist) the
check will fail since a stat is performed on the target (semantics of
stat(2)).

Signed-off-by: Brian Goff <cpuguy83@gmail.com>