tests: Do not unlazy the build!
Some things in here are a bit hacky.
Tests can modify the filesystem (even cause packages to be installed),
but we don't want those changes to be in the result we give to buildkit.
But if we don't include the test state in the build result then the
tests won't be executed... so this tries to get around the issue by
derving new a new state from the test state that *should* have the same
filesystem as the original, pre-test state.

This makes it so the dap debugger works when the spec has tests.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>