Don't unlazy the build
This breaks some assumptions for clients and breaks "docker buildx dap
build", which we are relying on for debugging work.

Before the change this is also technically solving multiple times, even
though the subsequent solves are cached.
The main culprit of unlazying the build is the test runner, which needs
to run commands and inspect filesystem state to return errors.

With this change the test runner is now executed as part of the normal
LLB graph.
All checks occur inside a buildkit container.
The frontend now has an extra set of subcommands for executing certain
checks. The frontend binary gets mounted into the test container and it
executes those subcommands to perform the various checks.

In order to preserve source mapping each test check is run in separate
exec ops.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>