tests: protect spec mutation with mutex
This prevents memory corruption due to spec mutation in the withSpec
test helper.
This is needed in case of multiple calls to `withSpec` with the same
spec pointer in parallel tests.
The race detector also throws an error here.

This fixes a potential easy to stuble upon gotcha within the test suite.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>